requirements.txt
panda3d
DirectFolderBrowser
BetterDirectGui
